Added max retry failed tuples Added move failed tuples Minor fixes

+++ src/spouts/files_mails.py
@@ -36,10 +36,12 @@ class FilesMailSpout(AbstractSpout):
 
         self.queue = Queue.PriorityQueue()
         self.queue_tail = set()
+        self.queue_fail = list()
         self.count = 1
         self.load_mails()
-        self._tuple_sleep = float(self.conf["tuple.sleep"])
-        self._waiting_sleep = float(self.conf["waiting.sleep"])
+        self.tuple_sleep = float(self.conf["tuple.sleep"])
+        self.waiting_sleep = float(self.conf["waiting.sleep"])
+        self.max_retry = int(self.conf["max.retry"])
 
     def load_mails(self):
         """This function load mails in a priority queue. """
@@ -47,26 +49,17 @@ class FilesMailSpout(AbstractSpout):
         self.log("Loading new mails for spout")
 
         mailboxes = self.conf['mailboxes']
-
         for k, v in mailboxes.iteritems():
-
             if not os.path.exists(v['path_mails']):
                 raise ImproperlyConfigured(
                     "Mail path '{}' does NOT exist".format(v['path_mails'])
                 )
 
-            all_mails = set(
-                glob.glob(
-                    os.path.join(
-                        v['path_mails'],
-                        '{}'.format(v['files_pattern']),
-                    )
-                )
-            )
-            new_mails = all_mails - self.queue_tail
+            all_mails = set(glob.glob(os.path.join(
+                v['path_mails'], '{}'.format(v['files_pattern']))))
 
             # put new mails in queue
-            for mail in new_mails:
+            for mail in (all_mails - self.queue_tail):
                 self.queue_tail.add(mail)
                 self.queue.put(
                     MailItem(
@@ -94,7 +87,7 @@ class FilesMailSpout(AbstractSpout):
                     ],
                     tup_id=mail.filename,
                 )
-                time.sleep(self._tuple_sleep)
+                time.sleep(self.tuple_sleep)
 
             # put new mails in priority queue
             else:
@@ -107,17 +100,25 @@ class FilesMailSpout(AbstractSpout):
         else:
             # Wait for new mails
             self.log("Queue mails is empty")
-            time.sleep(self._waiting_sleep)
+            time.sleep(self.waiting_sleep)
             self.load_mails()
 
     def ack(self, tup_id):
         """Acknowledge tup_id, that is the path_mail. """
 
         self.queue.task_done()
+        failed_tup = False
 
         try:
             # Remove from tail analyzed mail
             self.queue_tail.remove(tup_id)
+
+            # Remove from queue_tail
+            if tup_id in self.queue_fail:
+                failed_tup = True
+
+            self.queue_fail = [i for i in self.queue_fail if i != tup_id]
+
             self.log("Mails to process: {}".format(len(self.queue_tail)))
         except KeyError:
             pass
@@ -125,16 +126,20 @@ class FilesMailSpout(AbstractSpout):
         # Mails post processing
         what = self.conf["post_processing"]["what"].lower()
 
-        if what == "remove":
+        if what == "remove" and not failed_tup:
             # Delete mail if exists
             if os.path.exists(tup_id):
                 os.remove(tup_id)
 
         else:
-            where = self.conf["post_processing"]["where"]
+            if failed_tup:
+                where = self.conf["post_processing"]["where.failed"]
+            else:
+                where = self.conf["post_processing"]["where"]
+
             if not where:
                 raise ImproperlyConfigured(
-                    "Path where in '{}' is NOT configurated".format(
+                    "where or where.failed in '{}' is NOT configurated".format(
                         self.spouts_conf
                     )
                 )
@@ -153,6 +158,19 @@ class FilesMailSpout(AbstractSpout):
         # If tuple fail the mail remains on disk, self.queue is empty but
         # self.queue_tail contains all failed tuples
 
-        # Returns in queue
-        self.queue_tail.remove(tup_id)
-        self.log("Tuple '{}' failed. Now in queue".format(tup_id), "warning")
+        if self.queue_fail.count(tup_id) < self.max_retry:
+
+            # Returns in queue
+            self.queue_tail.remove(tup_id)
+
+            # Add in fail queue
+            self.queue_fail.append(tup_id)
+            self.log(
+                "Tuple '{}' failed. Now in queue".format(tup_id), "warning"
+            )
+        else:
+            self.log(
+                "Tuple {} failed for {} times".format(tup_id, self.max_retry)
+            )
+            # Remove from topology
+            self.ack(tup_id)
